backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-multipart
requests
psycopg2-binary
python-dotenv
